Migrate group API to new C API directory
This commit is just moving source code and setting up the necessary
changes to CMakeLists.txt. There are no changes to the implementation at
this stage.
